Bugfix. See below:
Don't deregister _roomRosterHandler when we receive a presence stanza
containing an error.

For example, the error might be an auth error. If the user then joins again
with the correct password, the same room is used and while the other handlers
(for presence and message) were being remembered, the roster handler wasn't
(because of this bug).